Move randomNonce method to String extension
